Covid-19 Data for Colombia & South Africa

This data contains 35 variables for Covid stats in Colombia and South Africa from January 2020.

covid_colsa

Format

An object of class grouped_df (inherits from tbl_df, tbl, data.frame) with 542 rows and 35 columns.

Source

The data comes from Guidotti, E., Ardia, D., (2020), "COVID-19 Data Hub", Journal of Open Source Software 5(51):2376, doi: 10.21105/joss.02376, through the R package "COVID19" and filtered for the country Colombia and South Africa.

Details

id - Country Code

date - Date of observation

tests - Total Count of tests for a date

confirmed - Total Count of confirmed for a date

recovered - Total Count of recovered for a date

deaths - Total Count of deaths for a date

key_apple_mobility - Country Name